--- metadata-ingestion/src/datahub/sql_parsing/sql_utils.py
@@ -1,5 +1,0 @@
-from typing import Union
-
-import sqlglot
-
-DialectOrStr = Union[sqlglot.Dialect, str]
